feat(animation): remove legacy animation stack and complete Animatable-only migration
- remove legacy Animation, AnimationManager, widget animation mixin, and easing module
- remove App legacy animation API and frame-stepping path
- decouple navigator transition handle typing from legacy widget animation helper
- remove legacy-only animation tests and update architecture docs
- verify with pytest, mypy, and flake8

Closes #20
